+++ native/src/rendering.cpp
@@ -44,8 +44,8 @@ void calcPoint(std::pair<int, int>  c, RenderingContext* rc)
 {
     rc->pointCount++;
 
-	double tx = c.first/ (rc->tileDivisor);
-	double ty = c.second / (rc->tileDivisor);
+	float tx = c.first/ (rc->tileDivisor);
+	float ty = c.second / (rc->tileDivisor);
 
     float dTileX = tx - rc->getLeft();
     float dTileY = ty - rc->getTop();
